Separate old and new models
