+++ corehq/apps/export/models/__init__.py
@@ -1,0 +1,12 @@
+from .deprecated import QuestionMeta, FormQuestionSchema
+from .new import (
+    ExportItem,
+    ExportColumn,
+    ExportRow,
+    ExportableItems,
+    TableConfiguration,
+    RepeatGroup,
+    ScalarItem,
+    Option,
+    MultipleChoiceItem,
+)
